// src/types.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  PENGURUS = 'PENGURUS',
  ANGGOTA = 'ANGGOTA',
}
// ...
export interface PermissionRequest {
  id: string;
  userId: string;
  eventId: string;
  reasonType: 'SICK' | 'FAMILY' | 'DUTY' | 'OTHER';
  reasonDetail: string;
  status: 'PENDING' | 'APPROVED' | 'REJECTED';
  timestamp: string;
  orgId?: string;
}